eth/fetcher: add metadata validation in tx announcement (ethereum#33378)  (#69)
* triedb/pathdb: use copy instead of append to reduce memory alloc (#33044)

* eth/fetcher: add metadata validation in tx announcement (#33378)

This PR fixes the bug reported in #33365.

The impact of the bug is not catastrophic. After a transaction is
ultimately fetched, validation and propagation will be performed based
on the fetched body, and any response with a mismatched type is treated
as a protocol violation. An attacker could only waste the limited
portion of victim’s bandwidth at most.

However, the reasons for submitting this PR are as follows

1. Fetching a transaction announced with an arbitrary type is a weird
behavior.

2. It aligns with efforts such as EIP-8077 and #33119 to make the
fetcher smarter and reduce bandwidth waste.

Regarding the `FilterType` function, it could potentially be implemented
by modifying the Filter function's parameter itself, but I wasn’t sure
whether changing that function is acceptable, so I left it as is.

* core/stateless: cap witness depth metrics buckets (#33389)

* cmd/utils: fix DeveloperFlag handling when set to false (#33379)

geth --dev=false now correctly respects the false value, instead of
incorrectly enabling UseLightweightKDF.

---------

Co-authored-by: Delweng <delweng@gmail.com>
Co-authored-by: Bosul Mun <bsbs8645@snu.ac.kr>
Co-authored-by: yyhrnk <yaroslavnerest@gmail.com>
Co-authored-by: Rizky Ikwan <rizzikwann@gmail.com>